First part of Blueprint implementation
